+++ src/types/chat_event.rs
@@ -12,11 +12,11 @@ pub struct ChatEvent {
   #[serde(rename(serialize = "@type", deserialize = "@type"))]
   td_name: String,
   /// Chat event identifier
-  id: i64,
+  id: isize,
   /// Point in time (Unix timestamp) when the event happened
-  date: i32,
+  date: i64,
   /// Identifier of the user who performed the action that triggered the event
-  user_id: i32,
+  user_id: i64,
   /// Action performed by the user
   action: ChatEventAction,
   
@@ -37,11 +37,11 @@ impl ChatEvent {
     RTDChatEventBuilder { inner }
   }
 
-  pub fn id(&self) -> i64 { self.id }
+  pub fn id(&self) -> isize { self.id }
 
-  pub fn date(&self) -> i32 { self.date }
+  pub fn date(&self) -> i64 { self.date }
 
-  pub fn user_id(&self) -> i32 { self.user_id }
+  pub fn user_id(&self) -> i64 { self.user_id }
 
   pub fn action(&self) -> &ChatEventAction { &self.action }
 
@@ -56,19 +56,19 @@ impl RTDChatEventBuilder {
   pub fn build(&self) -> ChatEvent { self.inner.clone() }
 
    
-  pub fn id(&mut self, id: i64) -> &mut Self {
+  pub fn id(&mut self, id: isize) -> &mut Self {
     self.inner.id = id;
     self
   }
 
    
-  pub fn date(&mut self, date: i32) -> &mut Self {
+  pub fn date(&mut self, date: i64) -> &mut Self {
     self.inner.date = date;
     self
   }
 
    
-  pub fn user_id(&mut self, user_id: i32) -> &mut Self {
+  pub fn user_id(&mut self, user_id: i64) -> &mut Self {
     self.inner.user_id = user_id;
     self
   }
