+++ src/types/chat_member.rs
@@ -12,11 +12,11 @@ pub struct ChatMember {
   #[serde(rename(serialize = "@type", deserialize = "@type"))]
   td_name: String,
   /// User identifier of the chat member
-  user_id: i32,
+  user_id: i64,
   /// Identifier of a user that invited/promoted/banned this member in the chat; 0 if unknown
-  inviter_user_id: i32,
+  inviter_user_id: i64,
   /// Point in time (Unix timestamp) when the user joined a chat
-  joined_chat_date: i32,
+  joined_chat_date: i64,
   /// Status of the member in the chat
   status: ChatMemberStatus,
   /// If the user is a bot, information about the bot; may be null. Can be null even for a bot if the bot is not a chat member
@@ -39,11 +39,11 @@ impl ChatMember {
     RTDChatMemberBuilder { inner }
   }
 
-  pub fn user_id(&self) -> i32 { self.user_id }
+  pub fn user_id(&self) -> i64 { self.user_id }
 
-  pub fn inviter_user_id(&self) -> i32 { self.inviter_user_id }
+  pub fn inviter_user_id(&self) -> i64 { self.inviter_user_id }
 
-  pub fn joined_chat_date(&self) -> i32 { self.joined_chat_date }
+  pub fn joined_chat_date(&self) -> i64 { self.joined_chat_date }
 
   pub fn status(&self) -> &ChatMemberStatus { &self.status }
 
@@ -60,19 +60,19 @@ impl RTDChatMemberBuilder {
   pub fn build(&self) -> ChatMember { self.inner.clone() }
 
    
-  pub fn user_id(&mut self, user_id: i32) -> &mut Self {
+  pub fn user_id(&mut self, user_id: i64) -> &mut Self {
     self.inner.user_id = user_id;
     self
   }
 
    
-  pub fn inviter_user_id(&mut self, inviter_user_id: i32) -> &mut Self {
+  pub fn inviter_user_id(&mut self, inviter_user_id: i64) -> &mut Self {
     self.inner.inviter_user_id = inviter_user_id;
     self
   }
 
    
-  pub fn joined_chat_date(&mut self, joined_chat_date: i32) -> &mut Self {
+  pub fn joined_chat_date(&mut self, joined_chat_date: i64) -> &mut Self {
     self.inner.joined_chat_date = joined_chat_date;
     self
   }
